move some stuff to common
